+++ app/web/src/services/email.service.ts
@@ -1,6 +1,5 @@
 import { Inject, Service } from 'typedi';
 import { ApiService } from 'src/services/api.service';
-import { NewsDto } from '@vdtn359/news-models';
 
 @Service()
 export class EmailService {