+++ app/web/src/services/news.service.ts
@@ -5,6 +5,7 @@ import {
 	TrackingService,
 	TrackingType,
 } from 'src/utils/tracking/tracking.service';
+import { logEvent } from 'src/utils/tracking/analytics';
 
 export type NewsSearchDto = NewsDto & { score: number; sort: any[] };
 
@@ -16,13 +17,15 @@ export class NewsService {
 		private readonly trackingService: TrackingService
 	) {}
 
-	get(slug: string) {
+	get(slug: string, params) {
 		return this.apiService
 			.api()
-			.get(`/news/${slug}`)
+			.get(`/news/${slug}`, {
+				params,
+			})
 			.then(({ data }) => data)
 			.catch((e) => {
-				if (e.response.status === 404) {
+				if (e.response?.status === 404) {
 					return null;
 				}
 				throw e;
@@ -124,6 +127,12 @@ export class NewsService {
 				type: TrackingType.UNLIKEID,
 			});
 		}
+		logEvent({
+			value: rating,
+			category: 'news',
+			action: 'rating',
+			label: id,
+		});
 		return this.apiService
 			.api()
 			.post('/rating', {